Move offer util.
