--- src/main/scala/mesosphere/marathon/core/task/update/TaskUpdateEffect.scala
@@ -1,14 +1,0 @@
-package mesosphere.marathon
-package core.task.update
-
-import mesosphere.marathon.core.task.Task
-
-object TaskUpdateEffect {
-  case object Noop extends TaskUpdateEffect
-  case class Update(newState: Task) extends TaskUpdateEffect
-
-  case class Failure(cause: Throwable) extends TaskUpdateEffect
-  object Failure {
-    def apply(message: String): Failure = Failure(new RuntimeException(message))
-  }
-}
